+++ android/src/main/java/com/marianhello/bgloc/LocationService.java
@@ -465,7 +465,7 @@ public class LocationService extends Service {
             int responseCode;
 
             try {
-                responseCode = HttpPostService.postFile(url, jsonLocations, config.getHttpHeaders());
+                responseCode = HttpPostService.postJSON(url, jsonLocations, config.getHttpHeaders());
             } catch (Exception e) {
                 hasConnectivity = isNetworkAvailable();
                 log.warn("Error while posting locations: {}", e.getMessage());