+++ android/src/main/java/com/marianhello/react/BackgroundGeolocationPackage.java
@@ -30,6 +30,6 @@ public class BackgroundGeolocationPackage implements ReactPackage {
 
   @Override
   public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
-      return Arrays.asList();
+      return Collections.emptyList();
   }
 }