+++ android/src/main/java/com/marianhello/utils/Convert.java
@@ -7,8 +7,11 @@ package com.marianhello.utils;
 
 public class Convert {
 
-  public static Integer safeLongToInt(Long l) {
-    if (l == null) return null;
-    return l.intValue();
+  public static int safeLongToInt(long l) {
+    if (l < Integer.MIN_VALUE || l > Integer.MAX_VALUE) {
+      throw new IllegalArgumentException
+        (l + " cannot be cast to int without changing its value.");
+    }
+    return (int) l;
   }
 }