+++ CHANGES.md
@@ -1,8 +1,16 @@
 ## Changelog
 
-### [0.2.0-alpha.2] - unreleased
+### [0.2.0-alpha.2] - 2016-08-31
 #### Fixed
 - fix config not persisted
+- tmp fix Android time long to int conversion
+
+#### Added
+- Android isLocationEnabled
+- Android showAppSettings
+- Android showLocationSettings
+- Android getLocations
+- Android getConfig
 
 ### [0.2.0-alpha.1] - 2016-08-17
 #### Changed